requirements.txt
fastapi
uvicorn
jinja2
beautifulsoup4
aiohttp
discord.py

